Do a better job of validating post length
This still doesn't correctly handle the shorter "effective lengths" for handles
and URLs. And it doesn't handle different per-server toot lengths. But it does
catch posts that become too long because the length of the CW is counted as part
of the length of the post body, and gives a validation error rather than
throwing an exception.