Update userinfo on settings page
There are cases where your session['user'] dictionary can be out of date, like
if you change your avatar. In that case, brutaldon will display a broken image
instead of your avatar in various places. This change updates your settings info
when you go to the settings page, in addition to as it normally does when you
log in.

Updating on every request would be possible, but it's another round trip to the
instance, and I don't want the performance hit.