Visibility of replies now more restrictive by default
Formerly, the visibility of replies defaulted to the visibility of the post
they were in reply to. Now they default to the more restrictive of the
visibility of the post they were in reply to, or your account's default
visibility. This should be less surprising for people with more restrictive
default visibility settings.